NOTIFICATION: Document TL280's encryption request message
There are 48 bytes in this message that seem to differ every time. There
is probably some info critical for understanding further encryption.

It's unlikely there's an actual AES key there. It also can't be RSA
public key (wouldn't fit 48 B).

Until we figure it out it's a real blocker for understanding /
implementing ITv2.